resolved Mr. dias' suggestions
