+++ commands/drive/teleop/swerve/XboxSwerveDrive.java
@@ -1,4 +1,4 @@
-package frc.team670.mustanglib.commands.drive.teleop;
+package frc.team670.mustanglib.commands.drive.teleop.swerve;
 
 import java.util.HashMap;
 import java.util.Map;
