+++ commands/drive/teleop/tank/XboxRocketLeague/ToggleIdleMode.java
@@ -1,4 +1,4 @@
-package frc.team670.mustanglib.commands.drive.teleop.XboxRocketLeague;
+package frc.team670.mustanglib.commands.drive.teleop.tank.XboxRocketLeague;
 
 import java.util.Map;
