+++ commands/drive/teleop/tank/XboxRocketLeague/XboxRobotOrientedDrive.java
@@ -1,4 +1,4 @@
-package frc.team670.mustanglib.commands.drive.teleop;
+package frc.team670.mustanglib.commands.drive.teleop.tank.XboxRocketLeague;
 
 import java.util.HashMap;
 import java.util.Map;
