--- swervelib/Mk3SwerveModuleHelper.java
@@ -1,509 +1,0 @@
-package frc.team670.mustanglib.swervelib;
-
-import frc.team670.mustanglib.swervelib.ctre.*;
-import frc.team670.mustanglib.swervelib.rev.*;
-import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardLayout;
-
-public final class Mk3SwerveModuleHelper {
-    private Mk3SwerveModuleHelper() {
-    }
-
-    private static DriveControllerFactory<?, Integer> getFalcon500DriveFactory(Mk3ModuleConfiguration configuration) {
-        return new Falcon500DriveControllerFactoryBuilder()
-                .withVoltageCompensation(configuration.getNominalVoltage())
-                .withCurrentLimit(configuration.getDriveCurrentLimit())
-                .build();
-    }
-
-    private static SteerControllerFactory<?, SteerConfiguration<CanCoderAbsoluteConfiguration>> getFalcon500SteerFactory(Mk3ModuleConfiguration configuration) {
-        return new Falcon500SteerControllerFactoryBuilder()
-                .withVoltageCompensation(configuration.getNominalVoltage())
-                .withPidConstants(0.2, 0.0, 0.1)
-                .withCurrentLimit(configuration.getSteerCurrentLimit())
-                .build(new CanCoderFactoryBuilder()
-                        .withReadingUpdatePeriod(100)
-                        .build());
-    }
-
-    private static DriveControllerFactory<?, Integer> getNeoDriveFactory(Mk3ModuleConfiguration configuration) {
-        return new NeoDriveControllerFactoryBuilder()
-                .withVoltageCompensation(configuration.getNominalVoltage())
-                .withCurrentLimit(configuration.getDriveCurrentLimit())
-                .build();
-    }
-
-    private static SteerControllerFactory<?, SteerConfiguration<CanCoderAbsoluteConfiguration>> getNeoSteerFactory(Mk3ModuleConfiguration configuration) {
-        return new NeoSteerControllerFactoryBuilder()
-                .withVoltageCompensation(configuration.getNominalVoltage())
-                .withPidConstants(1.0, 0.0, 0.1)
-                .withCurrentLimit(configuration.getSteerCurrentLimit())
-                .build(new CanCoderFactoryBuilder()
-                        .withReadingUpdatePeriod(100)
-                        .build());
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses Falcon 500s for driving and steering.
-     * Module information is displayed in the specified ShuffleBoard container.
-     *
-     * @param container        The container to display module information in.
-     * @param configuration    Module configuration parameters to use.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive Falcon 500.
-     * @param steerMotorPort   The CAN ID of the steer Falcon 500.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createFalcon500(
-            ShuffleboardLayout container,
-            Mk3ModuleConfiguration configuration,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return new SwerveModuleFactory<>(
-                gearRatio.getConfiguration(),
-                getFalcon500DriveFactory(configuration),
-                getFalcon500SteerFactory(configuration)
-        ).create(
-                container,
-                driveMotorPort,
-                new SteerConfiguration<>(
-                        steerMotorPort,
-                        new CanCoderAbsoluteConfiguration(steerEncoderPort, steerOffset)
-                )
-        );
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses Falcon 500s for driving and steering.
-     * Module information is displayed in the specified ShuffleBoard container.
-     *
-     * @param container        The container to display module information in.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive Falcon 500.
-     * @param steerMotorPort   The CAN ID of the steer Falcon 500.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createFalcon500(
-            ShuffleboardLayout container,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return createFalcon500(container, new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort, steerEncoderPort, steerOffset);
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses Falcon 500s for driving and steering.
-     *
-     * @param configuration    Module configuration parameters to use.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive Falcon 500.
-     * @param steerMotorPort   The CAN ID of the steer Falcon 500.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createFalcon500(
-            Mk3ModuleConfiguration configuration,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return new SwerveModuleFactory<>(
-                gearRatio.getConfiguration(),
-                getFalcon500DriveFactory(configuration),
-                getFalcon500SteerFactory(configuration)
-        ).create(
-                driveMotorPort,
-                new SteerConfiguration<>(
-                        steerMotorPort,
-                        new CanCoderAbsoluteConfiguration(steerEncoderPort, steerOffset)
-                )
-        );
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses Falcon 500s for driving and steering.
-     *
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive Falcon 500.
-     * @param steerMotorPort   The CAN ID of the steer Falcon 500.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createFalcon500(
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return createFalcon500(new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort, steerEncoderPort, steerOffset);
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses NEOs for driving and steering.
-     * Module information is displayed in the specified ShuffleBoard container.
-     *
-     * @param container        The container to display module information in.
-     * @param configuration    Module configuration parameters to use.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive NEO.
-     * @param steerMotorPort   The CAN ID of the steer NEO.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createNeo(
-            ShuffleboardLayout container,
-            Mk3ModuleConfiguration configuration,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return new SwerveModuleFactory<>(
-                gearRatio.getConfiguration(),
-                getNeoDriveFactory(configuration),
-                getNeoSteerFactory(configuration)
-        ).create(
-                container,
-                driveMotorPort,
-                new SteerConfiguration<>(
-                        steerMotorPort,
-                        new CanCoderAbsoluteConfiguration(steerEncoderPort, steerOffset)
-                )
-        );
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses NEOs for driving and steering.
-     * Module information is displayed in the specified ShuffleBoard container.
-     *
-     * @param container        The container to display module information in.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive NEO.
-     * @param steerMotorPort   The CAN ID of the steer NEO.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createNeo(
-            ShuffleboardLayout container,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return createNeo(container, new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort, steerEncoderPort, steerOffset);
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses NEOs for driving and steering.
-     *
-     * @param configuration    Module configuration parameters to use.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive NEO.
-     * @param steerMotorPort   The CAN ID of the steer NEO.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createNeo(
-            Mk3ModuleConfiguration configuration,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return new SwerveModuleFactory<>(
-                gearRatio.getConfiguration(),
-                getNeoDriveFactory(configuration),
-                getNeoSteerFactory(configuration)
-        ).create(
-                driveMotorPort,
-                new SteerConfiguration<>(
-                        steerMotorPort,
-                        new CanCoderAbsoluteConfiguration(steerEncoderPort, steerOffset)
-                )
-        );
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses NEOs for driving and steering.
-     *
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive NEO.
-     * @param steerMotorPort   The CAN ID of the steer NEO.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createNeo(
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return createNeo(new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort, steerEncoderPort, steerOffset);
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses a Falcon 500 for driving and a NEO for steering.
-     * Module information is displayed in the specified ShuffleBoard container.
-     *
-     * @param container        The container to display module information in.
-     * @param configuration    Module configuration parameters to use.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive Falcon 500.
-     * @param steerMotorPort   The CAN ID of the steer NEO.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createFalcon500Neo(
-            ShuffleboardLayout container,
-            Mk3ModuleConfiguration configuration,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return new SwerveModuleFactory<>(
-                gearRatio.getConfiguration(),
-                getFalcon500DriveFactory(configuration),
-                getNeoSteerFactory(configuration)
-        ).create(
-                container,
-                driveMotorPort,
-                new SteerConfiguration<>(
-                        steerMotorPort,
-                        new CanCoderAbsoluteConfiguration(steerEncoderPort, steerOffset)
-                )
-        );
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses a Falcon 500 for driving and a NEO for steering.
-     * Module information is displayed in the specified ShuffleBoard container.
-     *
-     * @param container        The container to display module information in.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive Falcon 500.
-     * @param steerMotorPort   The CAN ID of the steer NEO.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createFalcon500Neo(
-            ShuffleboardLayout container,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return createFalcon500Neo(container, new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort, steerEncoderPort, steerOffset);
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses a Falcon 500 for driving and a NEO for steering.
-     *
-     * @param configuration    Module configuration parameters to use.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive Falcon 500.
-     * @param steerMotorPort   The CAN ID of the steer NEO.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createFalcon500Neo(
-            Mk3ModuleConfiguration configuration,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return new SwerveModuleFactory<>(
-                gearRatio.getConfiguration(),
-                getFalcon500DriveFactory(configuration),
-                getNeoSteerFactory(configuration)
-        ).create(
-                driveMotorPort,
-                new SteerConfiguration<>(
-                        steerMotorPort,
-                        new CanCoderAbsoluteConfiguration(steerEncoderPort, steerOffset)
-                )
-        );
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses a Falcon 500 for driving and a NEO for steering.
-     *
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive Falcon 500.
-     * @param steerMotorPort   The CAN ID of the steer NEO.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createFalcon500Neo(
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return createFalcon500Neo(new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort, steerEncoderPort, steerOffset);
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses a NEO for driving and a Falcon 500 for steering.
-     * Module information is displayed in the specified ShuffleBoard container.
-     *
-     * @param container        The container to display module information in.
-     * @param configuration    Module configuration parameters to use.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive NEO.
-     * @param steerMotorPort   The CAN ID of the steer Falcon 500.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createNeoFalcon500(
-            ShuffleboardLayout container,
-            Mk3ModuleConfiguration configuration,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return new SwerveModuleFactory<>(
-                gearRatio.getConfiguration(),
-                getNeoDriveFactory(configuration),
-                getFalcon500SteerFactory(configuration)
-        ).create(
-                container,
-                driveMotorPort,
-                new SteerConfiguration<>(
-                        steerMotorPort,
-                        new CanCoderAbsoluteConfiguration(steerEncoderPort, steerOffset)
-                )
-        );
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses a NEO for driving and a Falcon 500 for steering.
-     * Module information is displayed in the specified ShuffleBoard container.
-     *
-     * @param container        The container to display module information in.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive NEO.
-     * @param steerMotorPort   The CAN ID of the steer Falcon 500.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createNeoFalcon500(
-            ShuffleboardLayout container,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return createNeoFalcon500(container, new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort, steerEncoderPort, steerOffset);
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses a NEO for driving and a Falcon 500 for steering.
-     *
-     * @param configuration    Module configuration parameters to use.
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive NEO.
-     * @param steerMotorPort   The CAN ID of the steer Falcon 500.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createNeoFalcon500(
-            Mk3ModuleConfiguration configuration,
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return new SwerveModuleFactory<>(
-                gearRatio.getConfiguration(),
-                getNeoDriveFactory(configuration),
-                getFalcon500SteerFactory(configuration)
-        ).create(
-                driveMotorPort,
-                new SteerConfiguration<>(
-                        steerMotorPort,
-                        new CanCoderAbsoluteConfiguration(steerEncoderPort, steerOffset)
-                )
-        );
-    }
-
-    /**
-     * Creates a Mk3 swerve module that uses a NEO for driving and a Falcon 500 for steering.
-     *
-     * @param gearRatio        The gearing configuration the module is in.
-     * @param driveMotorPort   The CAN ID of the drive NEO.
-     * @param steerMotorPort   The CAN ID of the steer Falcon 500.
-     * @param steerEncoderPort The CAN ID of the steer CANCoder.
-     * @param steerOffset      The offset of the CANCoder in radians.
-     * @return The configured swerve module.
-     */
-    public static SwerveModule createNeoFalcon500(
-            GearRatio gearRatio,
-            int driveMotorPort,
-            int steerMotorPort,
-            int steerEncoderPort,
-            double steerOffset
-    ) {
-        return createNeoFalcon500(new Mk3ModuleConfiguration(), gearRatio, driveMotorPort, steerMotorPort, steerEncoderPort, steerOffset);
-    }
-
-    public enum GearRatio {
-        /**
-         * Mk3 swerve in the standard gear configuration.
-         */
-        STANDARD(SdsModuleConfigurations.MK3_STANDARD),
-        /**
-         * Mk3 swerve in the fast gear configuration.
-         */
-        FAST(SdsModuleConfigurations.MK3_FAST);
-
-        private final ModuleConfiguration configuration;
-
-        GearRatio(ModuleConfiguration configuration) {
-            this.configuration = configuration;
-        }
-
-        public ModuleConfiguration getConfiguration() {
-            return configuration;
-        }
-    }
-}
